+++ VStancer.Client/VStancerConfig.cs
@@ -9,9 +9,9 @@
         public long Timer { get; set; }
         public int ToggleMenuControl { get; set; }
         public float FloatStep { get; set; }
-        public NodeLimits FrontLimits { get; set; }
-        public NodeLimits RearLimits { get; set; }
-        public Extra Extra { get; set; }
+        public ConfigNodeLimits FrontLimits { get; set; }
+        public ConfigNodeLimits RearLimits { get; set; }
+        public ConfigExtra Extra { get; set; }
 
         public VStancerConfig()
         {
@@ -22,21 +22,21 @@
             Timer = 1000;
             ToggleMenuControl = 167;
             FloatStep = 0.01f;
-            FrontLimits = new NodeLimits { PositionX = 0.25f, RotationY = 0.20f };
-            RearLimits = new NodeLimits { PositionX = 0.25f, RotationY = 0.20f };
+            FrontLimits = new ConfigNodeLimits { PositionX = 0.25f, RotationY = 0.20f };
+            RearLimits = new ConfigNodeLimits { PositionX = 0.25f, RotationY = 0.20f };
             
-            Extra = new Extra
+            Extra = new ConfigExtra
             {
                 EnableExtra = true,
                 WheelSize = 0.2f,
                 WheelWidth = 0.2f,
-                FrontWheelModSizeNodeLimit = new WheelModSizeNodeLimit
+                FrontWheelModSizeNodeLimit = new ConfigWheelModSizeNodeLimit
                 {
                     TireColliderScaleX = 0.1f,
                     TireColliderScaleYZ = 0.1f,
                     RimColliderScaleYZ = 0.1f,
                 },
-                RearWheelModSizeNodeLimit = new WheelModSizeNodeLimit
+                RearWheelModSizeNodeLimit = new ConfigWheelModSizeNodeLimit
                 {
                     TireColliderScaleX = 0.1f,
                     TireColliderScaleYZ = 0.1f,
@@ -46,25 +46,25 @@
         }
     }
 
-    public struct NodeLimits
+    public struct ConfigNodeLimits
     {
         public float PositionX { get; set; }
         public float RotationY { get; set; }
     }
 
-    public struct WheelModSizeNodeLimit
+    public struct ConfigWheelModSizeNodeLimit
     {
         public float TireColliderScaleX { get; set; }
         public float TireColliderScaleYZ { get; set; }
         public float RimColliderScaleYZ { get; set; }
     }
 
-    public struct Extra
+    public struct ConfigExtra
     {
         public bool EnableExtra { get; set; }
         public float WheelSize { get; set; }
         public float WheelWidth { get; set; }
-        public WheelModSizeNodeLimit FrontWheelModSizeNodeLimit { get; set; }
-        public WheelModSizeNodeLimit RearWheelModSizeNodeLimit { get; set; }
+        public ConfigWheelModSizeNodeLimit FrontWheelModSizeNodeLimit { get; set; }
+        public ConfigWheelModSizeNodeLimit RearWheelModSizeNodeLimit { get; set; }
     }
 }
