+++ VStancer.Client/VStancerUtilities.cs
@@ -1,0 +1,22 @@
+﻿namespace VStancer.Client
+{
+    public static class VStancerUtilities
+    {
+        public const float Epsilon = 0.001f;
+
+        /// <summary>
+        /// Calculate the number of front wheels of a vehicle, starting from the number of all the wheels
+        /// </summary>
+        /// <param name="wheelsCount">The number of wheels of a such vehicle</param>
+        /// <returns></returns>
+        public static int CalculateFrontWheelsCount(int wheelsCount)
+        {
+            int _frontWheelsCount = wheelsCount / 2;
+
+            if (_frontWheelsCount % 2 != 0)
+                _frontWheelsCount -= 1;
+
+            return _frontWheelsCount;
+        }
+    }
+}
